+++ ui/src/components/list-view/QasListView.stories.js
@@ -54,6 +54,10 @@ export default {
       description: '[VuexStoreModule](https://github.com/bildvitta/vuex-store-module) entity.'
     },
 
+    filtersProps: {
+      description: 'Controls filters props from [QasFilters](https://asteroid.nave.dev/?path=/docs/components-filters--default)'
+    },
+
     noFilter: {
       description: 'Hide filters.'
     },